requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==1.10.15
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
